# Speed up creation of simple composite units