With this PR, the call that gets done in creating prefix units:
```
import astropy.units as u
%timeit u.CompositeUnit(1.e-9, [u.m], [1], _error_check=False)
# 100000 loops, best of 3: 6.77 -> 1.05 µs per loop
```
As this is done a lot during import, it reduces the `import astropy.units` time by 20% (only the units part, not including the `astropy` part).

EDIT: I carried on a bit with this. Now:
```
%timeit u.m / u.s**2
10000 loops, best of 3: 17.3 -> 8.3 µs per loop
```
Import time reduced from 0.118 to 0.088 s
(Measured by `python -X importtime -c "import astropy; import astropy.units"` and looking at just the `astropy.units` part).