@mhvk 2018-08-04
@adrn - would you have a chance to review? I ask in part since in #7697 changes were proposed to `units/utils.py` that are no longer needed if this goes in.