@mhvk 2018-08-04
I pushed a rebased version with more comments; am happy to revert two changes that seem to be perhaps more confusing than it is worth; just let me know.