@eteq 2018-10-30
Awesome!  I think this is backport-able, do you agree @astrofrog @bsipocz ?