@astrofrog 2018-10-30
@eteq - agree!